# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: ashim API
  version: 1.15.9
  description: |
    REST API for ashim, a self-hosted image processing platform with 30+ tools.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Session cookie** — Call `POST /api/auth/login` with username and password. The response includes a `token` field. Pass it as `Authorization: Bearer <token>` on subsequent requests.

    2. **API key** — Generate a key via the Settings UI or `POST /api/v1/api-keys`. Keys are prefixed with `si_`. Pass as `Authorization: Bearer si_...`.

    Endpoints marked with a lock icon require authentication. Admin-only endpoints are noted in their description.
  license:
    name: AGPL-3.0
    url: https://github.com/ashim-hq/ashim/blob/main/LICENSE
servers:
  - url: /
    description: Current instance
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from login or API key (prefixed with si_)
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid image format
...
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Authentication required
...
    ConflictError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 409
        error:
          type: string
          example: Conflict
        message:
          type: string
          example: Resource already exists
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tools/barcode-read:
    post:
      tags: [Tools]
      summary: Barcode reader
      description: Read QR codes and barcodes from an image.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file containing barcode or QR code
      responses:
        "200":
          description: Decoded barcode data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                  location:
                    type: object
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
...
  /api/v1/teams:
    get:
      tags: [Teams]
      summary: List teams
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
    post:
      tags: [Teams]
      summary: Create a team
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "409":
          description: Team name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
  /api/v1/teams/{id}:
    put:
      tags: [Teams]
      summary: Rename a team
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Team integer ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
    delete:
      tags: [Teams]
      summary: Delete a team
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Team integer ID
          schema:
            type: integer
      responses:
        "204":
          description: Team deleted
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"

  # ─── Branding ─────────────────────────────────────────────────────────────
...